feat(exchange-rates): BCV fetch endpoint, status flow alignment, and tests
- Add POST /api/exchange-rates/bcv-fetch for manual BCV sync
- Add Draft-only validation to PUT /api/exchange-rates/{id}
- Add ExchangeRateService tests + bolivar payment tests (27 total)
- Rename SignalR events to match status-flow spec
- Fix DateTime.Kind=Utc for BCV scraper (prevents Npgsql errors)
- Fix Quartz cron expressions to 7-field format
- Change NewsletterDto.Status from enum to int in API response
- Add docs/status-flow.md with complete state machine spec

+++ AutoNex/DTOs/ExchangeRates/MappingExtensions.cs
@@ -10,7 +10,7 @@ public static class MappingExtensions
         PublishedAt = n.PublishedAt,
         ValueDate = n.ValueDate,
         Observations = n.Observations,
-        Status = n.Status,
+        Status = (int)n.Status,
         ExchangeRates = n.ExchangeRates.Select(r => new ExchangeRateDto
         {
             Id = r.Id,
